Add solution
This solution also assumes it's copacetic to use the python set
method to convert the input list. I do this to create a
de-deduplicated data structure with constant lookup time.
The same could be acheived instead with one more iteration through
the input list to generate a dict, which would still be O(n)