Major overhaul to deal with non adjacent char
There were some more iterations and test cases between the last
commit and this one, but I didn't record here every step as I
worked through it on leetcode

First, I came across a test that invalidated my assumption about only
needing to check adjacent characters, so I overhauled the approach
to check all permutations. This was brute force and contained a
nested loop, which I knew would be slow, but it passed all the
tests initially provided.

But on submit, it failed with Time Limit Exceeded on input of a
couple of Very long strings of almost all identical characters,
where the only difference was in the last two characters

This led me to realize that the Number of differences in the input
strings matter, and that we could significantly reduce runtime by
only paying attention to the differences between strings

This solution passes